+++ backend/package.json
@@ -12,7 +12,7 @@
   "dependencies": {
     "@ericblade/quagga2": "^1.8.2",
     "@types/passport": "^1.0.12",
-    "axios": "^1.3.4",
+    "axios": "^1.4.0",
     "bcrypt": "^5.0.0",
     "bcryptjs": "^2.4.3",
     "body-parser": "^1.19.0",
